Issue #99: Refactors: Moved page support modules (`CommentsPage`, `Compare`, `Match`, `ProcessDetailPage`, `ProcessesList`) to `modules/pages` folder.
