+++ bw_matchbox/assets/modules/pages/Compare/CompareRowClick.js
