+++ bw_matchbox/assets/modules/pages/Compare/CompareRowsHelpers.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import * as CommonHelpers from '../common/CommonHelpers.js';
+import * as CommonHelpers from '../../common/CommonHelpers.js';
 
 import { CompareRowClick } from './CompareRowClick.js';
