+++ bw_matchbox/assets/templates/comments.html
@@ -81,10 +81,10 @@
 <script type="module">
 
   // Core module
-  import { CommentsPage } from '/assets/modules/CommentsPage/CommentsPage.js';
+  import { CommentsPage } from '/assets/modules/pages/CommentsPage/CommentsPage.js';
 
   /** Global variables for compare tables feature (via global variable `sharedParams`)...
-   * @type {TSharedParams} - See `bw_matchbox/assets/modules/CommentsPage/@types/TSharedParams.d.ts`
+   * @type {TSharedParams} - See `bw_matchbox/assets/modules/pages/CommentsPage/@types/TSharedParams.d.ts`
    */
   const sharedParams = {
     base_url: '{{ base_url }}',
