+++ bw_matchbox/assets/templates/compare.html
@@ -69,7 +69,7 @@
 <script type="module">
 
   // Core module
-  import { CompareCore } from '/assets/modules/Compare/CompareCore.js';
+  import { CompareCore } from '/assets/modules/pages/Compare/CompareCore.js';
 
   // Global variables for compare tables feature (via global variable `sharedParams`)...
   const sharedParams = {
