+++ bw_matchbox/assets/templates/index.html
@@ -115,7 +115,7 @@
 <script type="module">
 
   // Core module
-  import { ProcessesList } from '/assets/modules/ProcessesList/ProcessesList.js';
+  import { ProcessesList } from '/assets/modules/pages/ProcessesList/ProcessesList.js';
 
   // Global variables for compare tables feature (via global variable `sharedParams`)...
   const sharedParams = {
