+++ bw_matchbox/assets/templates/match.html
@@ -40,7 +40,7 @@
 <script type="module">
 
   // Core module
-  import { Match } from '/assets/modules/Match/Match.js';
+  import { Match } from '/assets/modules/pages/Match/Match.js';
 
   // Global variables (via global variable `sharedParams`)...
   const sharedParams = {
