+++ bw_matchbox/assets/templates/process_detail.html
@@ -206,7 +206,7 @@
 <script type="module">
 
   // Core module
-  import { ProcessDetailPage } from '/assets/modules/ProcessDetailPage/ProcessDetailPage.js';
+  import { ProcessDetailPage } from '/assets/modules/pages/ProcessDetailPage/ProcessDetailPage.js';
 
   // Global variables for compare tables feature (via global variable `sharedParams`)...
   const sharedParams = {
